requirements.txt
anyio==3.5.0
asgiref==3.5.0
autopep8==1.6.0
bidict==0.21.4
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.4
fastapi==0.75.0
h11==0.13.0
idna==3.3
numpy==1.21.5
pandas==1.3.5
pycodestyle==2.8.0
pydantic==1.9.0
PyJWT==2.3.0
python-dateutil==2.8.2
python-dotenv==0.19.2
python-engineio==4.3.1
python-socketio==5.5.2
pytz==2021.3
pyupbit==0.2.23
requests==2.27.1
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
tabulate==0.8.9
toml==0.10.2
typing_extensions==4.1.1
urllib3==1.26.8
uvicorn==0.17.5
